Fix defaultOpenValue not aligned with step-size
It is possible for a user to enter a time that does not comply with
the defined steps (hourStep-, minuteStep-, secondStep property) when
no defaultOpenValue is supplied as a property. When defaultOpenValue is
not set, the current time is taken and thus a time not aligned with the
stepsize can appear in the input field.
Fix: The function toNearestValidTime() takes the defaultOpenValue, the valid
options for hour, minute and second and calculates the valid timestamp
that is nearest to the given defaultOpenValue.